+++ cmd/update.go
@@ -16,12 +16,16 @@ var updateCmd = &cobra.Command{
 	Use:   "update",
 	Short: "Update mangal",
 	Run: func(cmd *cobra.Command, args []string) {
+		cmd.Printf("%s %s\n", icon.Get(icon.Progress), "Fetching latest version...")
+
 		latestVersion, err := updater.LatestVersion()
 		handleErr(err)
 
-		if false && constant.Version >= latestVersion {
+		if constant.Version >= latestVersion {
 			cmd.Printf("%s %s\n", icon.Get(icon.Success), style.Green("You are using the latest version"))
 			return
+		} else {
+			cmd.Printf("%s New version is available: %s\n", icon.Get(icon.Success), latestVersion)
 		}
 
 		err = updater.Update()